+++ src/Main.py
@@ -1,4 +1,4 @@
-import src.FinanceAutomator as FinanceAutomator
+import FinanceAutomator as FinanceAutomator
 
 """
 Run this file in order to input transactional data into your Google Sheet.
@@ -13,18 +13,17 @@ If you want to inject all data of a folder
 Args:
     folder (str): Name of folder with CSV data files in there
 """
-file_path = './checkingsData/WFjune2022.csv'
+file_path = './checkingsData/WFaugust2022.csv'
 google_sheet = "Finance Tracker 2"
-sheet = "June 2022"
+sheet = "August 2022"
 folder = "checkingsData"
 
 
 
 if __name__ == "__main__":
-    # fa = Finance_Automator.Finance_Automator(google_sheet)
-    # fa.connect_to_google()
-    # # fa.inject_one_file_to_sheet(sheet, file_path)
+    fa = FinanceAutomator.FinanceAutomator(google_sheet)
+    fa.connect_to_google()
+    fa.inject_one_file_to_sheet(sheet, file_path)
     # fa.inject_all_data(folder)
     # file_path = "./venmoData/venmojune2022.csv"
-    import calendar
-    print(set([calendar.month_name[i].lower() for i in range(1, 13)]))
+
